# Log correct caller information when `SpanLogger` is used