@bboreham 2024-07-22
Is there a benchmark for this code?
Particularly interested whether we pay the stack-trace price for logs that are filtered out, eg debug logs.