@charleskorn 2024-07-24
@bboreham are you satisfied with the benchmarks above? If you have any ideas on how to make this less costly, I'd love to hear them.